+++ pkg/analyzer/lib/src/error/codes.g.dart
@@ -1256,7 +1256,7 @@ class CompileTimeErrorCode extends AnalyzerErrorCode {
   static const CompileTimeErrorCode EMPTY_MAP_PATTERN = CompileTimeErrorCode(
     'EMPTY_MAP_PATTERN',
     "A map pattern must have at least one entry.",
-    correctionMessage: "Try replacing it with an object pattern 'Map'.",
+    correctionMessage: "Try replacing it with an object pattern 'Map()'.",
   );
 
   static const CompileTimeErrorCode ENUM_CONSTANT_SAME_NAME_AS_ENCLOSING =
@@ -4335,7 +4335,7 @@ class CompileTimeErrorCode extends AnalyzerErrorCode {
       CompileTimeErrorCode(
     'REST_ELEMENT_IN_MAP_PATTERN',
     "A map pattern can't contain a rest pattern.",
-    correctionMessage: "Try removing the rest element.",
+    correctionMessage: "Try removing the rest pattern.",
   );
 
   ///  No parameters.