+++ CHANGELOG.md
@@ -282,7 +282,7 @@ using Dart version 2.19, before upgrading to Dart version 3.0.
 
 #### Pub
 
-- To preserve compatibility with null-safe code pre Dart 3 Pub will interpret a
+- To preserve compatibility with null-safe code pre Dart 3, Pub will interpret a
   language constraint indicating a language version of `2.12` or higher and an
   upper bound of `<3.0.0` as `<4.0.0`.
 
@@ -294,6 +294,9 @@ using Dart version 2.19, before upgrading to Dart version 3.0.
   2.1](https://www.rfc-editor.org/rfc/rfc6750#section-2.1). This means they must
   contain only the characters: `^[a-zA-Z0-9._~+/=-]+$`. Before a failure would
   happen when attempting to send the authorization header.
+- `dart pub get` and related commands will now by default also update the
+  dependencies in the `example` folder (if it exists). Use `--no-example` to
+  avoid this.
 
 ## 2.19.4 - 2023-03-08